soil sensor hecho

+++ final-project/soil.h
@@ -4,8 +4,8 @@
 #include "mbed.h"
 
 // Declaración de las variables globales
-extern volatile float soilData; // Variable externa para almacenar los datos
-extern Mutex soilMutex; // Mutex externo para proteger el acceso a soilData
+extern volatile int soilData; // Variable externa para almacenar los datos
+extern Mutex soilMutex; // Mutex externo para proteger el acceso a accelData
 
 // Declaración de la función de medición
 void soil_measurement();
